# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"cryptolab API","version":"0.1.0"},"paths":{"/decide/":{"post":{"tags":["decide"],"summary":"Decide","operationId":"decide_decide__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DecisionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DecisionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/train/":{"post":{"tags":["train"],"summary":"Train","operationId":"train_train__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/train/{task_id}":{"get":{"tags":["train"],"summary":"Get Train Task Status","operationId":"get_train_task_status_train__task_id__get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainTaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/backtest/":{"post":{"tags":["backtest"],"summary":"Backtest","operationId":"backtest_backtest__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BacktestRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BacktestResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/backtest/{task_id}":{"get":{"tags":["backtest"],"summary":"Get Backtest Task Status","operationId":"get_backtest_task_status_backtest__task_id__get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BacktestTaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/models/list":{"get":{"tags":["models"],"summary":"List Models","operationId":"list_models_models_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelListResponse"}}}}}}},"/models/info":{"post":{"tags":["models"],"summary":"Get Model Info","operationId":"get_model_info_models_info_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelInfoRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelInfoResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/explain/model/":{"post":{"tags":["explain"],"summary":"Explain","operationId":"explain_explain_model__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExplainModelRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExplainModelResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/explain/model/{task_id}":{"get":{"tags":["explain"],"summary":"Get Explanation","operationId":"get_explanation_explain_model__task_id__get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExplainModelTaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/explain/chart/":{"post":{"tags":["explain"],"summary":"Explain Chart","operationId":"explain_chart_explain_chart__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExplainChartRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExplainChartResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/explain/chart/{task_id}":{"get":{"tags":["explain"],"summary":"Get Chart Explanation","operationId":"get_chart_explanation_explain_chart__task_id__get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExplainChartTaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/score-chart/":{"post":{"tags":["score-chart"],"summary":"Score Chart","operationId":"score_chart_score_chart__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScoreChartRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScoreChartResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/score-chart/{task_id}":{"get":{"tags":["score-chart"],"summary":"Get Score Chart","operationId":"get_score_chart_score_chart__task_id__get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScoreChartTaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/data/list":{"get":{"tags":["data"],"summary":"List Coins","operationId":"list_coins_data_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CoinListResponse"}}}}}}},"/data/info":{"post":{"tags":["data"],"summary":"Get Coin Info","operationId":"get_coin_info_data_info_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CoinInfoRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CoinInfoResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/register":{"post":{"tags":["auth"],"summary":"Register","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/me":{"get":{"tags":["auth"],"summary":"Me","operationId":"me_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInfo"}}}}},"security":[{"HTTPBearer":[]}]}},"/watchlist":{"get":{"tags":["watchlist"],"summary":"Read Watchlist","operationId":"read_watchlist_watchlist_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchlistResponse"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["watchlist"],"summary":"Set Watchlist","operationId":"set_watchlist_watchlist_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchlistCreateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchlistResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}}},"components":{"schemas":{"BacktestRequest":{"properties":{"model_name":{"type":"string","title":"Model Name"},"param_name":{"type":"string","title":"Param Name"},"coin_symbol":{"type":"string","title":"Coin Symbol"},"timeframe":{"type":"integer","title":"Timeframe"},"start":{"type":"string","format":"date-time","title":"Start"},"end":{"type":"string","format":"date-time","title":"End"}},"type":"object","required":["model_name","param_name","coin_symbol","timeframe","start","end"],"title":"BacktestRequest"},"BacktestResponse":{"properties":{"task_id":{"type":"string","title":"Task Id"}},"type":"object","required":["task_id"],"title":"BacktestResponse"},"BacktestResult":{"properties":{"total_return":{"type":"number","title":"Total Return"},"win_rate":{"type":"number","title":"Win Rate"},"trade_count":{"type":"integer","title":"Trade Count"}},"type":"object","required":["total_return","win_rate","trade_count"],"title":"BacktestResult"},"BacktestTaskResponse":{"properties":{"task_id":{"type":"string","title":"Task Id"},"status":{"type":"string","title":"Status"},"results":{"anyOf":[{"$ref":"#/components/schemas/BacktestResult"},{"type":"null"}]}},"type":"object","required":["task_id","status"],"title":"BacktestTaskResponse"},"CoinInfoRequest":{"properties":{"coin_symbol":{"type":"string","title":"Coin Symbol"}},"type":"object","required":["coin_symbol"],"title":"CoinInfoRequest"},"CoinInfoResponse":{"properties":{"coin_symbol":{"type":"string","title":"Coin Symbol"},"available_start":{"type":"string","format":"date-time","title":"Available Start"},"available_end":{"type":"string","format":"date-time","title":"Available End"}},"type":"object","required":["coin_symbol","available_start","available_end"],"title":"CoinInfoResponse"},"CoinListResponse":{"properties":{"available_coin_symbols":{"items":{"type":"string"},"type":"array","title":"Available Coin Symbols"}},"type":"object","required":["available_coin_symbols"],"title":"CoinListResponse"},"DecisionRequest":{"properties":{"model_name":{"type":"string","title":"Model Name"},"param_name":{"type":"string","title":"Param Name"},"coin_symbol":{"type":"string","title":"Coin Symbol"},"timeframe":{"type":"integer","title":"Timeframe"},"inference_time":{"type":"string","format":"date-time","title":"Inference Time"},"cash_balance":{"type":"number","title":"Cash Balance"},"coin_balance":{"type":"number","title":"Coin Balance"}},"type":"object","required":["model_name","param_name","coin_symbol","timeframe","inference_time","cash_balance","coin_balance"],"title":"DecisionRequest"},"DecisionResponse":{"properties":{"action":{"type":"integer","title":"Action"},"amount":{"type":"number","title":"Amount"},"logit":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Logit"}},"type":"object","required":["action","amount"],"title":"DecisionResponse"},"ExplainChartRequest":{"properties":{"coin_symbol":{"type":"string","title":"Coin Symbol"},"timeframe":{"type":"integer","title":"Timeframe"},"inference_time":{"type":"string","format":"date-time","title":"Inference Time"},"start":{"type":"string","format":"date-time","title":"Start"},"end":{"type":"string","format":"date-time","title":"End"}},"type":"object","required":["coin_symbol","timeframe","inference_time","start","end"],"title":"ExplainChartRequest"},"ExplainChartResponse":{"properties":{"task_id":{"type":"string","title":"Task Id"}},"type":"object","required":["task_id"],"title":"ExplainChartResponse"},"ExplainChartResult":{"properties":{"similar_charts":{"items":{"$ref":"#/components/schemas/SimilarChartResult"},"type":"array","title":"Similar Charts"},"feature_values":{"additionalProperties":{"type":"number"},"type":"object","title":"Feature Values"},"explanation_text":{"type":"string","title":"Explanation Text"}},"type":"object","required":["similar_charts","feature_values","explanation_text"],"title":"ExplainChartResult"},"ExplainChartTaskResponse":{"properties":{"task_id":{"type":"string","title":"Task Id"},"status":{"type":"string","title":"Status"},"results":{"anyOf":[{"$ref":"#/components/schemas/ExplainChartResult"},{"type":"null"}]}},"type":"object","required":["task_id","status"],"title":"ExplainChartTaskResponse"},"ExplainModelRequest":{"properties":{"coin_symbol":{"type":"string","title":"Coin Symbol"},"timeframe":{"type":"integer","title":"Timeframe"},"inference_time":{"type":"string","format":"date-time","title":"Inference Time"}},"type":"object","required":["coin_symbol","timeframe","inference_time"],"title":"ExplainModelRequest"},"ExplainModelResponse":{"properties":{"task_id":{"type":"string","title":"Task Id"}},"type":"object","required":["task_id"],"title":"ExplainModelResponse"},"ExplainModelResult":{"properties":{"prediction_percentile":{"type":"number","title":"Prediction Percentile"},"recommendation":{"type":"string","title":"Recommendation"},"shap_values":{"additionalProperties":{"type":"number"},"type":"object","title":"Shap Values"},"feature_values":{"additionalProperties":{"type":"number"},"type":"object","title":"Feature Values"},"reference_charts":{"items":{"$ref":"#/components/schemas/ReferenceChartResult"},"type":"array","title":"Reference Charts"},"explanation_text":{"type":"string","title":"Explanation Text"}},"type":"object","required":["prediction_percentile","recommendation","shap_values","feature_values","reference_charts","explanation_text"],"title":"ExplainModelResult"},"ExplainModelTaskResponse":{"properties":{"task_id":{"type":"string","title":"Task Id"},"status":{"type":"string","title":"Status"},"results":{"anyOf":[{"$ref":"#/components/schemas/ExplainModelResult"},{"type":"null"}]}},"type":"object","required":["task_id","status"],"title":"ExplainModelTaskResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"email":{"type":"string","title":"Email"},"name":{"type":"string","title":"Name"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","name","password"],"title":"LoginRequest"},"LoginResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"expires_in":{"type":"integer","title":"Expires In","default":7200}},"type":"object","required":["access_token"],"title":"LoginResponse"},"ModelInfoRequest":{"properties":{"model_name":{"type":"string","title":"Model Name"}},"type":"object","required":["model_name"],"title":"ModelInfoRequest"},"ModelInfoResponse":{"properties":{"model_name":{"type":"string","title":"Model Name"},"hyperparam_schema":{"additionalProperties":true,"type":"object","title":"Hyperparam Schema"}},"type":"object","required":["model_name","hyperparam_schema"],"title":"ModelInfoResponse"},"ModelListResponse":{"properties":{"all_param_names":{"additionalProperties":{"items":{"type":"string"},"type":"array"},"type":"object","title":"All Param Names"}},"type":"object","required":["all_param_names"],"title":"ModelListResponse"},"ReferenceChartResult":{"properties":{"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"similarity":{"type":"number","title":"Similarity"}},"type":"object","required":["timestamp","similarity"],"title":"ReferenceChartResult"},"RegisterRequest":{"properties":{"email":{"type":"string","title":"Email"},"name":{"type":"string","maxLength":64,"minLength":1,"title":"Name"},"password":{"type":"string","maxLength":128,"minLength":8,"title":"Password"}},"type":"object","required":["email","name","password"],"title":"RegisterRequest"},"RegisterResponse":{"properties":{"user_id":{"type":"integer","title":"User Id"},"email":{"type":"string","title":"Email"},"name":{"type":"string","title":"Name"},"created_at":{"type":"string","title":"Created At"}},"type":"object","required":["user_id","email","name","created_at"],"title":"RegisterResponse"},"ScoreChartRequest":{"properties":{"coin_symbol":{"type":"string","title":"Coin Symbol"},"timeframe":{"type":"integer","title":"Timeframe"},"inference_time":{"type":"string","format":"date-time","title":"Inference Time"},"history_window":{"type":"integer","minimum":24.0,"title":"History Window","default":120}},"type":"object","required":["coin_symbol","timeframe","inference_time"],"title":"ScoreChartRequest"},"ScoreChartResponse":{"properties":{"task_id":{"type":"string","title":"Task Id"}},"type":"object","required":["task_id"],"title":"ScoreChartResponse"},"ScoreChartTaskResponse":{"properties":{"task_id":{"type":"string","title":"Task Id"},"status":{"type":"string","title":"Status"},"results":{"anyOf":[{"additionalProperties":{"$ref":"#/components/schemas/ScoreWithExplanation"},"type":"object"},{"type":"null"}],"title":"Results"}},"type":"object","required":["task_id","status"],"title":"ScoreChartTaskResponse"},"ScoreWithExplanation":{"properties":{"score":{"type":"number","title":"Score"},"explanation":{"type":"string","title":"Explanation"}},"type":"object","required":["score","explanation"],"title":"ScoreWithExplanation"},"SimilarChartResult":{"properties":{"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"distance":{"type":"number","title":"Distance"}},"type":"object","required":["timestamp","distance"],"title":"SimilarChartResult"},"TrainRequest":{"properties":{"model_name":{"type":"string","title":"Model Name"},"param_name":{"type":"string","title":"Param Name"},"coin_symbol":{"type":"string","title":"Coin Symbol"},"timeframe":{"type":"integer","title":"Timeframe"},"start":{"type":"string","format":"date-time","title":"Start"},"end":{"type":"string","format":"date-time","title":"End"},"hyperparams":{"additionalProperties":true,"type":"object","title":"Hyperparams"}},"type":"object","required":["model_name","param_name","coin_symbol","timeframe","start","end"],"title":"TrainRequest"},"TrainResponse":{"properties":{"task_id":{"type":"string","title":"Task Id"}},"type":"object","required":["task_id"],"title":"TrainResponse"},"TrainTaskResponse":{"properties":{"task_id":{"type":"string","title":"Task Id"},"status":{"type":"string","title":"Status"}},"type":"object","required":["task_id","status"],"title":"TrainTaskResponse"},"UserInfo":{"properties":{"user_id":{"type":"integer","title":"User Id"},"email":{"type":"string","title":"Email"},"name":{"type":"string","title":"Name"},"created_at":{"type":"string","title":"Created At"}},"type":"object","required":["user_id","email","name","created_at"],"title":"UserInfo"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WatchlistCreateRequest":{"properties":{"coin_symbols":{"items":{"type":"string"},"type":"array","maxItems":5,"minItems":5,"title":"Coin Symbols"}},"type":"object","required":["coin_symbols"],"title":"WatchlistCreateRequest"},"WatchlistResponse":{"properties":{"coin_symbols":{"items":{"type":"string"},"type":"array","title":"Coin Symbols"}},"type":"object","required":["coin_symbols"],"title":"WatchlistResponse"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}